feat: 뷰모델 코드 병합 및 네이밍 수정
