+++ clean-architecture-seminar/Domain /Models/UserModel.swift
@@ -20,36 +20,6 @@ struct UserModel: Equatable {
     let notifySetting: NotifySettingModel
     let createdAt: String
     let oauthAccount: OAuthAccountModel
-    
-    func update(
-        id: Int64? = nil,
-        username: String? = nil,
-        name: String? = nil,
-        isGeneralSignUp: Bool? = nil,
-        passwordUpdatedAt: String? = nil,
-        profileImageUrl: String? = nil,
-        phone: String? = nil,
-        profileVisibility: String? = nil,
-        locked: Bool? = nil,
-        notifySetting: NotifySettingModel? = nil,
-        createdAt: String? = nil,
-        oauthAccount: OAuthAccountModel? = nil
-    ) -> UserModel {
-        return UserModel(
-            id: id ?? self.id,
-            username: username ?? self.username,
-            name: name ?? self.name,
-            isGeneralSignUp: isGeneralSignUp ?? self.isGeneralSignUp,
-            passwordUpdatedAt: passwordUpdatedAt ?? self.passwordUpdatedAt,
-            profileImageUrl: profileImageUrl ?? self.profileImageUrl,
-            phone: phone ?? self.phone,
-            profileVisibility: profileVisibility ?? self.profileVisibility,
-            locked: locked ?? self.locked,
-            notifySetting: notifySetting ?? self.notifySetting,
-            createdAt: createdAt ?? self.createdAt,
-            oauthAccount: oauthAccount ?? self.oauthAccount
-        )
-    }
 }
 
 struct NotifySettingModel: Equatable {
