+++ clean-architecture-seminar/Infrastructure/Network/DataTransferService.swift
@@ -13,19 +13,7 @@ protocol DataTransferService {
 
 final class DefaultDataTransferService:  DataTransferService{
     
-//    private let networkService: NetworkService
-//    private let errorResolver: DataTransferErrorResolver
-//    private let errorLogger: DataTransferErrorLogger
-//    
-//    init(
-//        with networkService: NetworkService,
-//        errorResolver: DataTransferErrorResolver = DefaultDataTransferErrorResolver(),
-//        errorLogger: DataTransferErrorLogger = DefaultDataTransferErrorLogger()
-//    ) {
-//        self.networkService = networkService
-//        self.errorResolver = errorResolver
-//        self.errorLogger = errorLogger
-//    }
+
     func requset() {
         
     }
