+++ clean-architecture-seminar/Presentation/Scene/UserProfile/View/UserProfileView.swift
@@ -50,7 +50,7 @@ struct UserProfileView: View{
 
 /// `DefaultUserProfileViewModel`의 `userData` 변화를 관찰하여 UI가 자동으로 업데이트되도록 하는 클래스
 final class UserProfileViewModelWrapper: ObservableObject {
-    @Published var userData: UserModel
+    @Published var userData: UserProfileItemModel
     var viewModel: any UserProfileViewModel
     
     init(viewModel: any UserProfileViewModel) {
